Update langflow base prompts API utils
