+++ tests/test_loading.py
@@ -1,4 +1,5 @@
 import pytest
+
 from langflow.graph import Graph
 from langflow.graph.schema import RunOutputs
 from langflow.initial_setup.setup import load_starter_projects
@@ -36,7 +37,7 @@ def test_load_flow_from_json_object():
 def test_run_flow_from_json_object():
     """Test loading a flow from a json file and applying tweaks"""
     _, projects = zip(*load_starter_projects())
-    project = projects[0]
+    project = [project for project in projects if "Basic Prompting" in project["name"]][0]
     results = run_flow_from_json(project, input_value="test")
     assert results is not None
     assert all(isinstance(result, RunOutputs) for result in results)
